+++ src/wrapper_solvers.c
@@ -6,7 +6,7 @@
 #include "debug.h"
 #include "davidson.h"
 
-int sparse_eigensolve(double * vec, int dims, double* energy, 
+int sparse_eigensolve(double * vec, long long dims, double* energy, 
                       void (*matvec)(double*, double*, void*), 
                       const double * diagonal, double tol, int max_its, 
                       const char solver[], int davidson_keep,
